+++ src/Collision/ICollidable.java
@@ -2,5 +2,5 @@ package Collision;
 
 public interface ICollidable <T extends Particle> {
 	public void afterCollisionHandling (T p, T o);
-	public float getMass ();
+	public double getMass ();
 }
